reorganize folder structure & working on adding logo
