+++ src/assets/images/gatsby-icon.png
