+++ src/assets/images/smartea_logo_gray.png
